requirements.txt
google-api-python-client==2.179.0
python-dotenv==1.1.1
resend==2.13.0
Flask==3.0.0
filelock==3.13.1
pytz==2023.3
